Widen joint sweep limits to int so 500 survives
tick_from_lvl1_normalized took its sweep bounds as uint8_t, which caps at 255.
JOINT_MAX is 500 for every joint, so each upper bound was silently truncated to
244 before the tick was interpolated - every joint's travel was wrong at the top
of its range.

+++ walk.cpp
@@ -10,7 +10,7 @@ Walk walk;
 Adafruit_PWMServoDriver pwm = Adafruit_PWMServoDriver();
 void Walk::init(){};
 
-int  Walk::tick_from_lvl1_normalized(float norm, uint8_t jointSweepMin, uint8_t jointSweepMax){
+int  Walk::tick_from_lvl1_normalized(float norm, int jointSweepMin, int jointSweepMax){
    int tick=(jointSweepMax-jointSweepMin)*norm+jointSweepMin;
   return (int)tick;
